fix: query svm balance in BalanceAllocator
