+++ src/utils/TokenUtils.ts
@@ -1,10 +1,10 @@
 import { TOKEN_EQUIVALENCE_REMAPPING, TOKEN_SYMBOLS_MAP } from "@across-protocol/constants";
-import { constants, utils } from "@across-protocol/sdk";
+import { constants, utils, arch } from "@across-protocol/sdk";
 import { CONTRACT_ADDRESSES } from "../common";
-import { BigNumberish } from "./BNUtils";
+import { BigNumberish, BigNumber } from "./BNUtils";
 import { formatUnits } from "./SDKUtils";
 import { isDefined } from "./TypeGuards";
-import { Address, toAddressType, EvmAddress } from "./";
+import { Address, toAddressType, EvmAddress, SvmAddress, SVMProvider, toBN } from "./";
 
 const { ZERO_ADDRESS } = constants;
 
@@ -53,4 +53,49 @@ export function getWrappedNativeTokenAddress(chainId: number): Address {
 export function formatUnitsForToken(symbol: string, amount: BigNumberish): string {
   const decimals = (TOKEN_SYMBOLS_MAP[symbol]?.decimals as number) ?? 18;
   return formatUnits(amount, decimals);
+}
+
+/**
+ * Format the given amount of tokens to the correct number of decimals for the given token symbol.
+ * @param provider An SVM provider.
+ * @param tokenMint The address of the Solana token.
+ * @param walletAddress The address of the wallet to query.
+ * @returns The wallet's balance of the input token.
+ */
+export async function getSolanaTokenBalance(
+  provider: SVMProvider,
+  tokenMint: SvmAddress,
+  walletAddress: SvmAddress
+): Promise<BigNumber> {
+  // Convert addresses to the correct format for SVM provider
+  const ownerPubkey = arch.svm.toAddress(walletAddress);
+  const mintPubkey = arch.svm.toAddress(tokenMint);
+
+  // Get token accounts owned by the wallet for this specific mint
+  const tokenAccountsByOwner = await provider
+    .getTokenAccountsByOwner(
+      ownerPubkey,
+      {
+        mint: mintPubkey,
+      },
+      {
+        encoding: "jsonParsed",
+      }
+    )
+    .send();
+
+  const response = tokenAccountsByOwner;
+  if (!response.value || response.value.length === 0) {
+    // No token account found for this mint, balance is 0
+    return toBN(0);
+  }
+
+  // For SPL tokens, there should typically be only one token account per mint per owner
+  // Sum all balances in case there are multiple accounts (rare but possible)
+  const totalBalance = response.value.reduce((acc, accountInfo) => {
+    const balance = accountInfo?.account?.data?.parsed?.info?.tokenAmount?.amount;
+    return balance ? acc.add(toBN(balance)) : acc;
+  }, toBN(0));
+
+  return totalBalance;
 }
